disc price from decimal to float

+++ db/migrate/20220926215509_change_price_column_in_discs.rb
@@ -1,7 +1,7 @@
 class ChangePriceColumnInDiscs < ActiveRecord::Migration[7.0]
 
   def up
-    change_column :discs, :price, :decimal, :precision => 2
+    change_column :discs, :price, :float, :precision => 2
   end
   def down
     change_column :discs, :price, :decimal
